Add connectionstatus_test.go.
