+++ scrape/startupprocedure_test.go
@@ -1,10 +1,8 @@
 package scrape
 
 import (
-	"os"
 	"testing"
 
-	"github.com/PuerkitoBio/goquery"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -41,19 +39,4 @@ func TestScrapeStartupProcedure(t *testing.T) {
 	actual := scrapeStartupProcedure(doc)
 	assert.NotNil(t, actual)
 	assert.Equal(t, expected, actual)
-}
-
-func getConnectionStatusDocumentFromTestFile(t *testing.T) *goquery.Document {
-	filePath := "../testdata/sb8200/cmconnectionstatus.html"
-	fileReader, err := os.Open(filePath)
-	if err != nil {
-		t.Fatalf("unable to open file for reading: [%s]", filePath)
-	}
-
-	doc, err := goquery.NewDocumentFromReader(fileReader)
-	if err != nil {
-		t.Fatalf("unable to generate goquery document from file: [%s]", filePath)
-	}
-
-	return doc
 }
